+++ src/CoreBundle/Twig/Extension/StatusExtension.php
@@ -18,6 +18,8 @@ use Twig\TwigFilter;
 /**
  * @author Hugo Briand <briand@ekino.com>
  * @author Titouan Galopin <galopintitouan@gmail.com>
+ *
+ * @deprecated Since version 3.x, to be removed in 4.0.
  */
 class StatusExtension extends AbstractExtension
 {
