+++ src/CoreBundle/Twig/TokenParser/DeprecatedTemplateTokenParser.php
@@ -11,30 +11,17 @@
 
 namespace Sonata\CoreBundle\Twig\TokenParser;
 
-use Sonata\CoreBundle\Twig\Node\DeprecatedTemplateNode;
-use Twig\Token;
-use Twig\TokenParser\AbstractTokenParser;
+@trigger_error(
+    'The '.__NAMESPACE__.'\DeprecatedTemplateTokenParser class is deprecated since version 3.x and will be removed in 4.0.'
+    .' Use Sonata\Twig\TokenParser\DeprecatedTemplateTokenParser instead.',
+    E_USER_DEPRECATED
+);
 
 /**
  * @author Marko Kunic <kunicmarko20@gmail.com>
+ *
+ * @deprecated Since version 3.x, to be removed in 4.0.
  */
-final class DeprecatedTemplateTokenParser extends AbstractTokenParser
+final class DeprecatedTemplateTokenParser extends \Sonata\Twig\TokenParser\DeprecatedTemplateTokenParser
 {
-    public function parse(Token $token)
-    {
-        if (!$this->parser->getStream()->test(Token::STRING_TYPE)) {
-            throw new \InvalidArgumentException('New template name is mandatory.');
-        }
-
-        $newTemplate = $this->parser->getExpressionParser()->parseExpression();
-
-        $this->parser->getStream()->expect(Token::BLOCK_END_TYPE);
-
-        return new DeprecatedTemplateNode($newTemplate, $token->getLine(), $this->getTag());
-    }
-
-    public function getTag()
-    {
-        return 'sonata_template_deprecate';
-    }
 }
