+++ tests/Twig/Extension/FlashMessageExtensionTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sonata\CoreBundle\Tests\Twig\Extension;
+namespace Sonata\Twig\Tests\Twig\Extension;
 
 use PHPUnit\Framework\TestCase;
 use Sonata\Twig\Extension\FlashMessageExtension;
