+++ tests/Twig/Node/TemplateBoxNodeTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sonata\CoreBundle\Tests\Twig\TokenParser;
+namespace Sonata\Twig\Tests\Twig\TokenParser;
 
 use Sonata\Twig\Node\TemplateBoxNode;
 use Symfony\Component\Translation\Formatter\MessageFormatterInterface;
