+++ tests/Twig/TokenParser/DeprecatedTemplateTokenParserTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sonata\CoreBundle\Tests\Twig\TokenParser;
+namespace Sonata\Twig\Tests\Twig\TokenParser;
 
 use PHPUnit\Framework\TestCase;
 use Sonata\Twig\Node\DeprecatedTemplateNode;
